Model/requirements.txt
fastapi
uvicorn
torch
torchvision
opencv-python-headless
python-dotenv
requests
numpy
pillow
